% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_em.R
\name{fct_em}
\alias{fct_em}
\title{Internal EM Algorithm}
\usage{
fct_em(x, y, k, lambda, clust_assign, lik_track, em_iter, verbose)
}
\arguments{
\item{x}{A matrix of predictors.}

\item{y}{A matrix of responses.}

\item{k}{The number of groups.}

\item{lambda}{A vector of penalization parameters.}

\item{clust_assign}{The current clustering assignment.}

\item{lik_track}{A vector storing the log-likelihood by iteration.}

\item{em_iter}{The maximum number of EM iterations.}

\item{verbose}{A boolean indicating whether to print to screen.}
}
\value{
A mixedLSR model.
}
\description{
Internal EM Algorithm
}
\keyword{internal}
