% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gamma.R
\name{fct_gamma}
\alias{fct_gamma}
\title{Internal Posterior Calculation}
\usage{
fct_gamma(
  x,
  y,
  k,
  N,
  clust_assign,
  pi_vec,
  lambda,
  alpha,
  beta,
  y_sparse,
  rank,
  max_rank
)
}
\arguments{
\item{x}{A matrix of predictors.}

\item{y}{A matrix of responses.}

\item{k}{The number of groups.}

\item{N}{The sample size.}

\item{clust_assign}{The current clustering assignment.}

\item{pi_vec}{A vector of mixing probabilities for each cluster label.}

\item{lambda}{A vector of penalization parameters.}

\item{alpha}{A positive constant DPP parameter.}

\item{beta}{A positive constant DPP parameter.}

\item{y_sparse}{Should Y coefficients be treated as sparse?}

\item{rank}{The rank, if known.}

\item{max_rank}{The maximum allowed rank.}
}
\value{
A list with the posterior, coefficients, and estimated covariance.
}
\description{
Internal Posterior Calculation
}
\keyword{internal}
