% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_initialize.R
\name{fct_initialize}
\alias{fct_initialize}
\title{Internal Partition Initialization Function}
\usage{
fct_initialize(k, N)
}
\arguments{
\item{k}{The number of groups.}

\item{N}{The sample size.}
}
\value{
A vector of assignments.
}
\description{
Internal Partition Initialization Function
}
\keyword{internal}
