% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_j_lik.R
\name{fct_j_lik}
\alias{fct_j_lik}
\title{Internal Likelihood Function}
\usage{
fct_j_lik(
  x,
  y,
  k,
  clust_assign,
  lambda,
  alpha = 2 * sqrt(3),
  beta = 1,
  y_sparse = TRUE,
  max_rank = 3,
  rank = NULL
)
}
\arguments{
\item{x}{A matrix of predictors.}

\item{y}{A matrix of responses.}

\item{k}{The number of groups.}

\item{clust_assign}{A vector of cluster labels.}

\item{lambda}{A vector of penalization parameters.}

\item{alpha}{A positive constant DPP parameter.}

\item{beta}{A positive constant DPP parameter.}

\item{y_sparse}{Should Y coefficients be treated as sparse?}

\item{max_rank}{The maximum allowed rank.}

\item{rank}{The rank, if known.}
}
\value{
The weighted log-likelihood
}
\description{
Internal Likelihood Function
}
\keyword{internal}
