% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_log_lik.R
\name{fct_log_lik}
\alias{fct_log_lik}
\title{Internal Log-Likelihood Function}
\usage{
fct_log_lik(mu_mat, sig_vec, y, N, m)
}
\arguments{
\item{mu_mat}{The mean matrix.}

\item{sig_vec}{A vector of sigma.}

\item{y}{The output matrix.}

\item{N}{The sample size.}

\item{m}{The number of y features.}
}
\value{
A posterior matrix.
}
\description{
Internal Log-Likelihood Function
}
\keyword{internal}
