% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_new_assign.R
\name{fct_new_assign}
\alias{fct_new_assign}
\title{Internal Perturb Function}
\usage{
fct_new_assign(assign, k, p)
}
\arguments{
\item{assign}{The current clustering assignments.}

\item{k}{The number of groups.}

\item{p}{The acceptance probability.}
}
\value{
A perturbed assignment.
}
\description{
Internal Perturb Function
}
\keyword{internal}
