% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pi_vec.R
\name{fct_pi_vec}
\alias{fct_pi_vec}
\title{Internal Pi Function}
\usage{
fct_pi_vec(clust_assign, k, N)
}
\arguments{
\item{clust_assign}{The current clustering assignment.}

\item{k}{The number of groups.}

\item{N}{The sample size.}
}
\value{
A mixing vector.
}
\description{
Internal Pi Function
}
\keyword{internal}
