% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rank.R
\name{fct_rank}
\alias{fct_rank}
\title{Internal Rank Estimation Function}
\usage{
fct_rank(x, y, sigma, eta)
}
\arguments{
\item{x}{A matrix of predictors.}

\item{y}{A matrix of responses.}

\item{sigma}{An estimated noise level.}

\item{eta}{A rank selection parameter.}
}
\value{
The estimated rank.
}
\description{
Internal Rank Estimation Function
}
\keyword{internal}
