% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_select_lambda.R
\name{fct_select_lambda}
\alias{fct_select_lambda}
\title{Internal Penalty Parameter Selection Function.}
\usage{
fct_select_lambda(
  x,
  y,
  k,
  clust_assign = NULL,
  initial = FALSE,
  type = "all",
  verbose
)
}
\arguments{
\item{x}{A matrix of predictors.}

\item{y}{A matrix of responses.}

\item{k}{The number of groups.}

\item{clust_assign}{The current clustering assignment.}

\item{initial}{An initial penalty parameter.}

\item{type}{A type.}

\item{verbose}{A boolean indicating whether to print to screen.}
}
\value{
A selected penalty parameter.
}
\description{
Internal Penalty Parameter Selection Function.
}
\keyword{internal}
