% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sigma.R
\name{fct_sigma}
\alias{fct_sigma}
\title{Internal Sigma Estimation Function}
\usage{
fct_sigma(y, N, m)
}
\arguments{
\item{y}{A matrix of responses.}

\item{N}{The sample size.}

\item{m}{The number of outcome variables.}
}
\value{
The estimated sigma.
}
\description{
Internal Sigma Estimation Function
}
\keyword{internal}
