% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_weighted_ll.R
\name{fct_weighted_ll}
\alias{fct_weighted_ll}
\title{Internal Weighted Log Likelihood Function}
\usage{
fct_weighted_ll(gamma)
}
\arguments{
\item{gamma}{A posterior matrix}
}
\value{
A weighted log likelihood vector
}
\description{
Internal Weighted Log Likelihood Function
}
\keyword{internal}
