% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lsr.R
\name{plot_lsr}
\alias{plot_lsr}
\title{Heatmap Plot of the mixedLSR Coefficient Matrices}
\usage{
plot_lsr(a, abs = TRUE)
}
\arguments{
\item{a}{A coefficient matrix from mixed_lsr model.}

\item{abs}{A boolean for taking the absolute value of the coefficient matrix.}
}
\value{
A ggplot2 heatmap of the coefficient matrix, separated by subgroup.
}
\description{
Heatmap Plot of the mixedLSR Coefficient Matrices
}
\examples{
simulate <- simulate_lsr()
plot_lsr(simulate$a)
}
