% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lsr.R
\name{simulate_lsr}
\alias{simulate_lsr}
\title{Simulate Heterogeneous, Low-Rank, and Sparse Data}
\usage{
simulate_lsr(
  N = 100,
  k = 2,
  p = 30,
  m = 35,
  b = 1,
  d = 20,
  h = 0.2,
  case = "independent"
)
}
\arguments{
\item{N}{The sample size, default = 100.}

\item{k}{The number of groups, default = 2.}

\item{p}{The number of predictor features, default = 30.}

\item{m}{The number of response features, default = 35.}

\item{b}{The signal-to-noise ratio, default = 1.}

\item{d}{The singular value, default = 20.}

\item{h}{The lower bound for the singular matrix simulation, default = 0.2.}

\item{case}{The covariance case, "independent" or "dependent", default = "independent".}
}
\value{
A list of simulation values, including x matrix, y matrix, coefficients and true clustering assignments.
}
\description{
Simulate Heterogeneous, Low-Rank, and Sparse Data
}
\examples{
simulate_lsr()
}
