% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sparse.R
\name{simulate_sparse}
\alias{simulate_sparse}
\title{Simulate Heterogeneous, Low-Rank Data with Varying Sparsity}
\usage{
simulate_sparse(k = 2, dense = 0.1)
}
\arguments{
\item{k}{The number of groups, default = 2.}

\item{dense}{The density ratio (must be greater than 0).}
}
\value{
A list of simulation values, including x matrix, y matrix, coefficients and true clustering assignments.
}
\description{
Simulate Heterogeneous, Low-Rank Data with Varying Sparsity
}
\examples{
simulate_lsr()
}
